requirements.txt
bs4==0.0.1
cchardet==2.1.7
Flask-Cors==3.0.10
Flask==2.0.3
gunicorn==20.0.4
lxml==4.8.0
PyNaCl==1.3.0
requests==2.27.1
virtualenv==20.4.2
PyGitHub==1.55
